refactor: change components name
