+++ src/components/pages/Home/index.stories.tsx
@@ -1,16 +1,13 @@
 import React from 'react';
-import { Provider } from 'react-redux';
 import { storiesOf } from '@storybook/react';
 import { StoryHeader } from '../../organisms/Header/index.stories';
-import Home from './container';
-import { store } from '../../../store';
+import Contact from '.';
 
 storiesOf('pages/Home', module)
-  .addDecorator(story => <Provider store={store}>{story()}</Provider>)
-  .add('default', () => <Home />)
+  .add('default', () => <Contact />)
   .add('with layout', () => (
     <div>
       <StoryHeader />
-      <Home />
+      <Contact />
     </div>
   ));
